requirements.txt
# LINE Bot
line-bot-sdk>=3.7.0
flask>=3.0.0

# 向量資料庫
chromadb>=0.5.0

# 文件處理
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.0

# AI/ML
groq>=0.4.0
sentence-transformers>=2.3.0
numpy>=1.26.0
Pillow>=10.2.0  # 圖像處理

# 工具
python-dotenv>=1.0.0
requests>=2.31.0
